gbn: make bumped value for processNACK correct
Previously, the bumped value for processNACK didn't accurately reflect
if the queue sequence base was actually bumped or not for the cases when
we didn't trigger a resend. This commit addresses that issue.

This has no real impact on the current implementation of the gbn conn,
but it might start having an effect if the implementation is modified
in the future.